Gave player stamina, added crystal count for ex zones, and fixed ui nav for pause menu
Player now has a stamina bar that decreases while sprinting and jumping. It is also displayed via a bar in the middle of the screen.

Ex zones now show how many crystals it has extracted, and how much space it has left before being disabled

Forgot to do the UI nav for the pause menu and the quit modal, but added that in. Had to make a workaround that isn't pretty for the resume button because it was double-inputting with the jump input when pressing "A" on the controller.