More bulk, see description
Finished the UI navigation for the controller the most part. Very messy, but it will have to do.

Added the sound system to most interactions

Created medium and hard levels with different extraction zones and enemies in them

Did some small UI updated like show the player's AD and the enemies health points